Update: Refactored 'app.js'
Refactored _app.js_ into './src/security' and 'src/utils' while taking into account the `1 routed
added = 2 NCLOCs` problem which is now solved with a dynamic requiring which would only need one
key/pair line added per new route.

fix #78
